requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3

# Database & Caching
pymongo==4.6.0
redis==5.0.1

# LLM & AI
langgraph==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38

# Performance & Production
slowapi==0.1.9
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2